refactor: create storage dir
