+++ src/utils.ts
@@ -2,7 +2,7 @@ import { ContainsArrayError, HTTPStore, openArray, openGroup, ZarrArray } from '
 import type { Group as ZarrGroup } from 'zarr';
 import type { AsyncStore, Store } from 'zarr/types/storage/types';
 import { Matrix4 } from '@math.gl/core/dist/esm';
-import { LRUCacheStore } from './lru-store';
+import { LRUCacheStore } from './storage/lru-store';
 
 export const MAX_CHANNELS = 6;
